modify

+++ velox/functions/prestosql/ArithmeticImpl.h
@@ -45,7 +45,7 @@ round(const TNum& number, const TDecimals& decimals = 0) {
       return number;
     }
   }
-  if (!std::isfinite(number)) {
+  if (!std::isfinite(static_cast<double>(number))) {
     return number;
   }
